+++ packages/node/src/core/providers/process-transactions.ts
@@ -2,7 +2,7 @@ import { ProviderState } from '../../types';
 import * as ethereum from '../ethereum';
 import * as logger from '../utils/logger';
 import * as nonces from '../transactions/nonces';
-import * as transactions from '../transactions/submitting';
+import * as transactions from '../ethereum/transactions';
 import * as state from './state';
 
 export async function processTransactions(initialState: ProviderState) {
